Add precalc stuff to bobs and cels. Clean up

+++ ScrollerMapper/Converters/Infos/OutputPathStepInfo.cs
@@ -4,8 +4,8 @@ namespace ScrollerMapper.Converters.Infos
     internal class OutputPathStepInfo
     {
         public byte FrameCount { get; set; }
-        public byte VelocityX { get; set; }
-        public byte VelocityY { get; set; }
+        public short VelocityX { get; set; }
+        public short VelocityY { get; set; }
 
         public override string ToString()
         {
